:sparkles: feat: fever와 normal 서비스 계층 분리
